+++ dbt/aus_personas/tests/assert_pgm_language_home_english_proficiency_probability_sums.sql
@@ -1,0 +1,13 @@
+with probability_sums as (
+    select
+        census_year,
+        sa2_code,
+        sex,
+        sum(probability_within_partition) as probability_sum
+    from {{ ref('mart_pgm__sa2_language_home_english_proficiency') }}
+    group by 1, 2, 3
+)
+
+select *
+from probability_sums
+where abs(probability_sum - 1.0) > 0.000001